backend/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.34.24
geoalchemy2==0.14.3
shapely==2.0.2
aiofiles==23.2.1
httpx==0.26.0
celery==5.3.6
python-dotenv==1.0.1
